Make error applicable on iOS
