+++ detox/test/index.ios.js
@@ -1,3 +1,4 @@
+import {LaunchArguments} from 'react-native-launch-arguments';
 import example from './src/app';
 
 import {
@@ -5,6 +6,10 @@ import {
 } from 'react-native';
 
 class exampleIos extends example {}
+
+if (LaunchArguments.value().simulateEarlyCrash) {
+  throw new Error('Simulating early crash');
+}
 
 console.disableYellowBox = true;
 AppRegistry.registerComponent('example', () => exampleIos);
